api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "uvicorn (>=0.41.0,<0.42.0)",
    "fastapi (>=0.134.0,<0.135.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-core (>=1.2.16,<2.0.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "socksio (>=1.0.0,<2.0.0)",
    "arize-phoenix-otel (>=0.10.0)",
    "openinference-instrumentation-langchain (>=0.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
